+++ packages/theme/src/prebuilt/ts/color-tokens.ts
@@ -18,7 +18,7 @@ export default {
 	'primary-stroke3': [
 		'background-thumb-brand',
 		'background-thumb-brand-active',
-		'stroke-focus-brand',
+		'stroke-focus',
 		'stroke-interactive-brand',
 		'stroke-surface-brand-strong',
 	],